moved module workflow to design package
