+++ src/compas_timber/design/__init__.py
@@ -1,0 +1,17 @@
+from .workflow import CategoryRule
+from .workflow import DirectRule
+from .workflow import JointRule
+from .workflow import TopologyRule
+from .workflow import JointDefinition
+from .workflow import FeatureDefinition
+from .workflow import DebugInfomation
+
+__all__ = [
+    "CategoryRule",
+    "DirectRule",
+    "JointRule",
+    "TopologyRule",
+    "JointDefinition",
+    "FeatureDefinition",
+    "DebugInfomation",
+]
