+++ src/compas_timber/ghpython/__init__.py
@@ -1,20 +1,8 @@
 from .ghcomponent_helpers import add_gh_param
 from .ghcomponent_helpers import clear_gh_params
 from .ghcomponent_helpers import manage_dynamic_params
-from .workflow import CategoryRule
-from .workflow import DebugInfomation
-from .workflow import DirectRule
-from .workflow import FeatureDefinition
-from .workflow import JointDefinition
-from .workflow import TopologyRule
 
 __all__ = [
-    "JointDefinition",
-    "CategoryRule",
-    "TopologyRule",
-    "DirectRule",
-    "FeatureDefinition",
-    "DebugInfomation",
     "clear_gh_params",
     "add_gh_param",
     "manage_dynamic_params",
