+++ src/compas_timber/ghpython/components/CT_Feature_Trim/code.py
@@ -3,7 +3,7 @@ from ghpythonlib.componentbase import executingcomponent as component
 from Grasshopper.Kernel.GH_RuntimeMessageLevel import Warning
 
 from compas_timber.elements import CutFeature
-from compas_timber.ghpython import FeatureDefinition
+from compas_timber.design import FeatureDefinition
 
 
 class TrimmingFeature(component):
